Refactor:move a bunch of the import utility functions to a separate module
