+++ src/components/FileImportExport.test.js
@@ -1,4 +1,4 @@
-import { isRoomNumberAsString, parseRow, parseRows } from './FileImportExport';
+import { isRoomNumberAsString, parseRow, parseRows } from './importUtils';
 
 describe('isRoomNumberAsString', () => {
   it('should return true for a valid number string', () => {
